requirements.txt
django==4.0.3
djangorestframework==3.13.1
pandas==1.4.2
wetterdienst==0.31.1
django-rest-knox==4.2.0
openpyxl==3.0.9
numpy==1.22.3
psycopg2==2.9.3